+++ src/components/Login.jsx
@@ -5,7 +5,7 @@ import '../style/Login.css';
 import { API_URL } from '../shared';
 import { useNavigate } from 'react-router-dom';
 
-const Login = () => {
+const Login = ( {setUser} ) => {
   const [isLogin, setIsLogin] = useState(true);
   const [showPassword, setShowPassword] = useState(false);
   const [currentSlide, setCurrentSlide] = useState(0);
@@ -68,36 +68,56 @@ const Login = () => {
     }));
   };
 
-  // Helper to switch between login and signup
+const handleSpotifyLogin = async () => {
+  try {
+    setLoading(true);
+    const response = await axios.get(`${API_URL}/auth/spotify/login-url`);
+    window.location.href = response.data.authUrl;
+  } catch (error) {
+    console.error('Spotify login error:', error.message);
+    setError("Failed to connect to Spotify");
+    setLoading(false);
+  }
+};
+
   const handleAuth = async (e) => {
     e.preventDefault();
     setError("");
     setLoading(true);
     try {
-      // accept either email or username
       const identifier = formData.email.trim();
       if (!identifier) throw new Error("Email or username is required");
+      
       if (isLogin) {
         // Login
         const response = await axios.post(`${API_URL}/auth/login`, {
           username: identifier,
           password: formData.password,
         }, { withCredentials: true });
+        
         if (response.data.error) throw new Error(response.data.error);
-        setError("");
-        // Redirect to /analytics after successful login
-        navigate('/analytics');
+        
+        if (response.data.token) {
+          localStorage.setItem('authToken', response.data.token);
+        }
+        
+        setUser(response.data.user); // Use the setUser prop
+        navigate('/dashboard/analytics'); // Navigate to dashboard
       } else {
         // Signup
         const response = await axios.post(`${API_URL}/auth/signup`, {
           username: identifier,
-          email: identifier,
           password: formData.password,
         }, { withCredentials: true });
+        
         if (response.data.error) throw new Error(response.data.error);
-        setError("");
-        // Redirect to login form after successful signup
-        setIsLogin(true);
+        
+        if (response.data.token) {
+          localStorage.setItem('authToken', response.data.token);
+        }
+        
+        setUser(response.data.user); 
+        navigate('/dashboard/analytics'); 
       }
     } catch (err) {
       setError(err.response?.data?.error || err.message);
@@ -161,7 +181,7 @@ const Login = () => {
 
           {/* OAuth Buttons */}
           <div className="oauth-section">
-            <button className="spotify-btn">
+            <button className="spotify-btn" onClick={handleSpotifyLogin} disabled={loading}>
               <svg className="spotify-icon" fill="currentColor" viewBox="0 0 24 24">
                 <path d="M12 0C5.4 0 0 5.4 0 12s5.4 12 12 12 12-5.4 12-12S18.66 0 12 0zm5.521 17.34c-.24.359-.66.48-1.021.24-2.82-1.74-6.36-2.101-10.561-1.141-.418.122-.779-.179-.899-.539-.12-.421.18-.78.54-.9 4.56-1.021 8.52-.6 11.64 1.32.42.18.479.659.301 1.02zm1.44-3.3c-.301.42-.841.6-1.262.3-3.239-1.98-8.159-2.58-11.939-1.38-.479.12-1.02-.12-1.14-.6-.12-.48.12-1.021.6-1.141C9.6 9.9 15 10.561 18.72 12.84c.361.181.54.78.241 1.2zm.12-3.36C15.24 8.4 8.82 8.16 5.16 9.301c-.6.179-1.2-.181-1.38-.721-.18-.601.18-1.2.72-1.381 4.26-1.26 11.28-1.02 15.721 1.621.539.3.719 1.02.42 1.56-.299.421-1.02.599-1.559.3z"/>
               </svg>
