--- src/pages/Signup.jsx
@@ -1,206 +1,0 @@
-import React, { useState } from "react";
-import { useNavigate, Link } from "react-router-dom";
-import { useAuth0 } from "@auth0/auth0-react";
-import axios from "axios";
-import { API_URL } from "../shared";
-import "../style/Signup.css";
-
-const Signup = ({ setUser }) => {
-  const [formData, setFormData] = useState({
-    username: "",
-    password: "",
-    confirmPassword: "",
-  });
-  const [errors, setErrors] = useState({});
-  const [isLoading, setIsLoading] = useState(false);
-  const navigate = useNavigate();
-
-  const { loginWithRedirect } = useAuth0();
-
-  const validateForm = () => {
-    const newErrors = {};
-
-    if (!formData.username) {
-      newErrors.username = "Username is required";
-    } else if (formData.username.length < 3 || formData.username.length > 20) {
-      newErrors.username = "Username must be between 3 and 20 characters";
-    } else if (!/^[a-zA-Z0-9_]+$/.test(formData.username)) {
-      newErrors.username = "Username can only contain letters, numbers, and underscores";
-    }
-
-    if (!formData.password) {
-      newErrors.password = "Password is required";
-    } else if (formData.password.length < 6) {
-      newErrors.password = "Password must be at least 6 characters";
-    }
-
-    if (!formData.confirmPassword) {
-      newErrors.confirmPassword = "Please confirm your password";
-    } else if (formData.password !== formData.confirmPassword) {
-      newErrors.confirmPassword = "Passwords do not match";
-    }
-
-    setErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!validateForm()) {
-      return;
-    }
-
-    setIsLoading(true);
-    setErrors({}); 
-
-    try {
-      const response = await axios.post(`${API_URL}/auth/signup`, {
-        username: formData.username,
-        password: formData.password,
-      }, {
-        withCredentials: true,
-      });
-
-      if (response.data.token) {
-        localStorage.setItem('authToken', response.data.token);
-      }
-
-      setUser(response.data.user);
-      navigate("/", { 
-        state: { 
-          message: "Account created successfully! Welcome!" 
-        } 
-      });
-
-    } catch (error) {
-      if (error.response?.status === 409) {
-        setErrors({ username: "This username is already taken. Please choose a different one." });
-      } else if (error.response?.data?.error) {
-        setErrors({ general: error.response.data.error });
-      } else if (error.response?.status === 400) {
-        setErrors({ general: "Please check your input and try again." });
-      } else {
-        setErrors({ general: "An error occurred during signup. Please try again." });
-      }
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleChange = (e) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: value,
-    }));
-
-    if (errors[name]) {
-      setErrors((prev) => ({
-        ...prev,
-        [name]: "",
-      }));
-    }
-
-    if (errors.general) {
-      setErrors((prev) => ({
-        ...prev,
-        general: "",
-      }));
-    }
-  };
-
-  const handleAuth0Signup = () => {
-    loginWithRedirect({
-      authorizationParams: {
-        screen_hint: "signup",
-        connection: 'Username-Password-Authentication',
-      },
-    });
-  };
-
-  return (
-    <div className="auth-container">
-      <div className="auth-form">
-        <h2>Sign Up</h2>
-
-        {errors.general && (
-          <div className="error-message">{errors.general}</div>
-        )}
-
-        <form onSubmit={handleSubmit}>
-          <div className="form-group">
-            <label htmlFor="username">Username:</label>
-            <input
-              type="text"
-              id="username"
-              name="username"
-              value={formData.username}
-              onChange={handleChange}
-              className={errors.username ? "error" : ""}
-              placeholder="Choose a username (3-20 characters)"
-            />
-            {errors.username && (
-              <span className="error-text">{errors.username}</span>
-            )}
-          </div>
-
-          <div className="form-group">
-            <label htmlFor="password">Password:</label>
-            <input
-              type="password"
-              id="password"
-              name="password"
-              value={formData.password}
-              onChange={handleChange}
-              className={errors.password ? "error" : ""}
-              placeholder="Enter a password (minimum 6 characters)"
-            />
-            {errors.password && (
-              <span className="error-text">{errors.password}</span>
-            )}
-          </div>
-
-          <div className="form-group">
-            <label htmlFor="confirmPassword">Confirm Password:</label>
-            <input
-              type="password"
-              id="confirmPassword"
-              name="confirmPassword"
-              value={formData.confirmPassword}
-              onChange={handleChange}
-              className={errors.confirmPassword ? "error" : ""}
-              placeholder="Confirm your password"
-            />
-            {errors.confirmPassword && (
-              <span className="error-text">{errors.confirmPassword}</span>
-            )}
-          </div>
-
-          <button type="submit" disabled={isLoading} className="submit-btn">
-            {isLoading ? "Creating Account..." : "Sign Up"}
-          </button>
-        </form>
-
-        <div className="auth-divider">
-          <span>or</span>
-        </div>
-
-        <button
-          type="button"
-          onClick={handleAuth0Signup}
-          className="auth0-login-btn"
-          disabled={isLoading}
-        >
-          Sign up with Auth0
-        </button>
-
-        <p className="auth-link">
-          Already have an account? <Link to="/login">Log in</Link>
-        </p>
-      </div>
-    </div>
-  );
-};
-
-export default Signup;
